+++ sdk/monitor/azquery/build.go
@@ -1,5 +1,5 @@
-//go:generate autorest ./autorest.md --rawjson-as-bytes
-//go:generate gofmt -w .
+//go:generate autorest ./autorest.md
+//go:generate goimports -w .
 
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License. See License.txt in the project root for license information.