+++ sdk/monitor/azquery/models_serde.go
@@ -8,9 +8,10 @@ package azquery
 import (
 	"encoding/json"
 	"fmt"
+	"reflect"
+
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore"
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore/to"
-	"reflect"
 )
 
 // MarshalJSON implements the json.Marshaller interface for type BatchQueryRequest.