+++ sdk/monitor/azquery/CHANGELOG.md
@@ -4,12 +4,15 @@
 
 ### Features Added
 * Added API Version support for `MetricsClient`. Users can now change the default API Version by setting `MetricsClientOptions.APIVersion`.
+* Added `AutoAdjustTimegrain`, `RollUpBy`, and `ValidateDimensions` options to `MetricsClient.QueryResource`
 
 ### Breaking Changes
+* Removed `MetricsBatchClient`
 
 ### Bugs Fixed
 
 ### Other Changes
+* Upgrade ARM metrics API version to `2024-02-01`
 
 ## 1.2.0-beta.1 (2023-11-16)